search: Modify bucket implementation for benchmark (#65)
Noticed that ~50% of the benchmark time was spent doing syscall.Open.
The filesystem bucket implementation we're using come from
thanos/objstore repository so I decided to copy it over and cache the
file descriptors so that we can run get_range calls in parallel without
getting throttled by the syscalls.

Signed-off-by: Jesus Vazquez <jesusvzpg@gmail.com>